pyproject.toml
[tool.poetry]
name = "metis"
version = "1.31.0"
description = "Card enrolment bridge API. Handles enrolment and unenrolment of payment cards"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "metis"}]

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "azure"
url = "https://pkgs.dev.azure.com/binkhq/_packaging/binkhq/pypi/simple/"
priority = "explicit"

[tool.poetry.dependencies]
python = "^3.11"
arrow = "^1.3.0"
celery = "^5.3.5"
lxml = "^4.9.3"
pika = "^1.3.2"
requests = "^2.31.0"
pyjwt = "^2.8.0"
sentry-sdk = "^1.36.0"
prometheus-client = "^0.19.0"
gunicorn = "^21.2.0"
azure-keyvault-secrets = "^4.7.0"
azure-identity = "^1.15.0"
loguru = "^0.7.2"
bink-logging-utils = { version="^1.4.0", source="azure" }
fastapi = "^0.109.2"
pydantic-settings = "^2.1.0"
uvicorn = "^0.27.0.post1"
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"
pytest = "^7.4.3"
pyfakefs = "^5.3.1"
isort = "^5.12.0"
black = "^23.11.0"
pytest-cov = "^4.1.0"
xenon = "^0.9.1"
httpretty = "^1.0.5, !=1.1.0, !=1.1.1, !=1.1.2, !=1.1.3, !=1.1.4"
flask-testing = "^0.8.1"
typer = "^0.9.0"
ipython = "^8.17.2"
refurb = "^1.24.0"
respx = "^0.20.2"
pytest-xdist = "^3.5.0"
mypy = "^1.8.0"
types-requests = "^2.31.0.20240125"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
disallow_untyped_defs = true
show_error_codes = true
warn_unused_ignores = true
ignore_missing_imports = true
disallow_incomplete_defs = true
warn_redundant_casts = true
# exclude = "alembic/versions/"
plugins = [
    "pydantic.mypy",
]

[tool.ruff]
line-length = 120
show-source = true
target-version = "py311"
select = [
    "F",    # Pyflakes (F)
    "E",
    "W",    # pycodestyle (E, W)
    "C90",  # mccabe (C90)
    # isort (I)
    # pydocstyle (D)
    "UP",   # pyupgrade (UP)
    "N",    # pep8-naming (N)
    "YTT",  # flake8-2020 (YTT)
    # "ANN",  # flake8-annotations (ANN)
    # "S",    # flake8-bandit (S)
    "BLE",  # flake8-blind-except (BLE)
    # flake8-boolean-trap (FBT)
    "B",    # flake8-bugbear (B)
    "A",    # flake8-builtins (A)
    "C4",   # flake8-comprehensions (C4)
    "T10",  # flake8-debugger (T10)
    # flake8-errmsg (EM)
    "ICN",  # flake8-import-conventions (ICN)
    "T20",  # flake8-print (T20)
    # "PT",   # flake8-pytest-style (PT)
    "Q",    # flake8-quotes (Q)
    # flake8-return (RET)
    "SIM",  # flake8-simplify (SIM)
    "TID",  # flake8-tidy-imports (TID)
    "ARG",  # flake8-unused-arguments (ARG)
    "DTZ",  # flake8-datetimez (DTZ)
    "ERA",  # eradicate (ERA) until we can remove commented out code this has to stay deactivated
    # pandas-vet (PD)
    # pygrep-hooks (PGH)
    "PLC",
    "PLE",
    "PLR",
    "PLW",  # Pylint (PLC, PLE, PLR, PLW)
    "RUF",  # Ruff-specific rules (RUF)
]

ignore = [
    "BLE001",
    "PLR2004",
    "DTZ005", # FIXME: all the datetimes in this project are datetime.now() with no tz specified is this correct?
    "N818",
    "PLR0915", # Too many statements
    "RUF012", # FIXME: This is a temporary fix to allow us to use the new Ruff rules without breaking the build
    "RUF013", # FIXME: This is a temporary fix to allow us to use the new Ruff rules without breaking the build
]

[tool.ruff.per-file-ignores]
"tests/*" = ["T201", "PLR0913"]
"metis/prometheus/*" = ["T201"]
"metis/card_router.py" = ["PLR0915"]

[tool.refurb]
ignore = [
    "FURB184",  # some chaining suggestions don't make sense
]
